+++ consensus/src/pipeline/virtual_processor/tests_virtual.rs
@@ -1,160 +1,16 @@
-﻿use crate::{consensus::test_consensus::TestConsensus, model::services::reachability::ReachabilityService};
+﻿use super::tests_util::{new_miner_data, OnetimeTxSelector, TestContext};
+use crate::{consensus::test_consensus::TestConsensus, model::services::reachability::ReachabilityService};
 use kaspa_consensus_core::{
     BlockHashSet,
     api::ConsensusApi,
-    block::{Block, BlockTemplate, MutableBlock, TemplateBuildMode, TemplateTransactionSelector},
-    blockhash,
+    block::TemplateBuildMode,
     blockstatus::BlockStatus,
-    coinbase::MinerData,
     config::{
         ConfigBuilder,
         params::{ForkActivation, MAINNET_PARAMS},
     },
     constants::{BLOCK_VERSION, TOCCATA_BLOCK_VERSION},
-    tx::{ScriptPublicKey, ScriptVec, Transaction},
 };
-use kaspa_hashes::Hash;
-use std::{collections::VecDeque, thread::JoinHandle};
-
-struct OnetimeTxSelector {
-    txs: Option<Vec<Transaction>>,
-}
-
-impl OnetimeTxSelector {
-    fn new(txs: Vec<Transaction>) -> Self {
-        Self { txs: Some(txs) }
-    }
-}
-
-impl TemplateTransactionSelector for OnetimeTxSelector {
-    fn select_transactions(&mut self) -> Vec<Transaction> {
-        self.txs.take().unwrap()
-    }
-
-    fn reject_selection(&mut self, _tx_id: kaspa_consensus_core::tx::TransactionId) {
-        unimplemented!()
-    }
-
-    fn is_successful(&self) -> bool {
-        true
-    }
-}
-
-struct TestContext {
-    consensus: TestConsensus,
-    join_handles: Vec<JoinHandle<()>>,
-    miner_data: MinerData,
-    simulated_time: u64,
-    current_templates: VecDeque<BlockTemplate>,
-    current_tips: BlockHashSet,
-}
-
-impl Drop for TestContext {
-    fn drop(&mut self) {
-        self.consensus.shutdown(std::mem::take(&mut self.join_handles));
-    }
-}
-
-impl TestContext {
-    fn new(consensus: TestConsensus) -> Self {
-        let join_handles = consensus.init();
-        let genesis_hash = consensus.params().genesis.hash;
-        let simulated_time = consensus.params().genesis.timestamp;
-        Self {
-            consensus,
-            join_handles,
-            miner_data: new_miner_data(),
-            simulated_time,
-            current_templates: Default::default(),
-            current_tips: BlockHashSet::from_iter([genesis_hash]),
-        }
-    }
-
-    pub fn build_block_template_row(&mut self, nonces: impl Iterator<Item = usize>) -> &mut Self {
-        for nonce in nonces {
-            self.simulated_time += self.consensus.params().target_time_per_block();
-            self.current_templates.push_back(self.build_block_template(nonce as u64, self.simulated_time));
-        }
-        self
-    }
-
-    pub fn assert_row_parents(&mut self) -> &mut Self {
-        for t in self.current_templates.iter() {
-            assert_eq!(self.current_tips, BlockHashSet::from_iter(t.block.header.direct_parents().iter().copied()));
-        }
-        self
-    }
-
-    pub async fn validate_and_insert_row(&mut self) -> &mut Self {
-        self.current_tips.clear();
-        while let Some(t) = self.current_templates.pop_front() {
-            self.current_tips.insert(t.block.header.hash);
-            self.validate_and_insert_block(t.block.to_immutable()).await;
-        }
-        self
-    }
-
-    pub async fn build_and_insert_disqualified_chain(&mut self, mut parents: Vec<Hash>, len: usize) -> Hash {
-        // The chain will be disqualified since build_block_with_parents builds utxo-invalid blocks
-        for _ in 0..len {
-            self.simulated_time += self.consensus.params().target_time_per_block();
-            let b = self.build_block_with_parents(parents, 0, self.simulated_time);
-            parents = vec![b.header.hash];
-            self.validate_and_insert_block(b.to_immutable()).await;
-        }
-        parents[0]
-    }
-
-    pub fn build_block_template(&self, nonce: u64, timestamp: u64) -> BlockTemplate {
-        let mut t = self
-            .consensus
-            .build_block_template(
-                self.miner_data.clone(),
-                Box::new(OnetimeTxSelector::new(Default::default())),
-                TemplateBuildMode::Standard,
-            )
-            .unwrap();
-        t.block.header.timestamp = timestamp;
-        t.block.header.nonce = nonce;
-        t.block.header.finalize();
-        t
-    }
-
-    pub fn build_block_with_parents(&self, parents: Vec<Hash>, nonce: u64, timestamp: u64) -> MutableBlock {
-        let mut b = self.consensus.build_block_with_parents_and_transactions(blockhash::NONE, parents, Default::default());
-        b.header.timestamp = timestamp;
-        b.header.nonce = nonce;
-        b.header.finalize(); // This overrides the NONE hash we passed earlier with the actual hash
-        b
-    }
-
-    pub async fn validate_and_insert_block(&mut self, block: Block) -> &mut Self {
-        let status = self.consensus.validate_and_insert_block(block).virtual_state_task.await.unwrap();
-        assert!(status.has_block_body());
-        self
-    }
-
-    pub fn assert_tips(&mut self) -> &mut Self {
-        assert_eq!(BlockHashSet::from_iter(self.consensus.get_tips().into_iter()), self.current_tips);
-        self
-    }
-
-    pub fn assert_tips_num(&mut self, expected_num: usize) -> &mut Self {
-        assert_eq!(BlockHashSet::from_iter(self.consensus.get_tips().into_iter()).len(), expected_num);
-        self
-    }
-
-    pub fn assert_virtual_parents_subset(&mut self) -> &mut Self {
-        assert!(self.consensus.get_virtual_parents().is_subset(&self.current_tips));
-        self
-    }
-
-    pub fn assert_valid_utxo_tip(&mut self) -> &mut Self {
-        // Assert that at least one body tip was resolved with valid UTXO
-        assert!(self.consensus.body_tips().iter().copied().any(|h| self.consensus.block_status(h) == BlockStatus::StatusUTXOValid));
-        self
-    }
-}
 
 #[tokio::test]
 async fn template_mining_sanity_test() {
@@ -333,14 +189,6 @@ async fn double_search_disqualified_test() {
         ctx.build_block_template_row(0..1).validate_and_insert_row().await.assert_valid_utxo_tip();
     }
     ctx.assert_tips_num(1);
-}
-
-fn new_miner_data() -> MinerData {
-    let secp = secp256k1::Secp256k1::new();
-    let mut rng = rand::thread_rng();
-    let (_sk, pk) = secp.generate_keypair(&mut rng);
-    let script = ScriptVec::from_slice(&pk.serialize());
-    MinerData::new(ScriptPublicKey::new(0, script), vec![])
 }
 
 fn inactivity_shortcut_config() -> kaspa_consensus_core::config::Config {
